+++ bibliotecaComunitaria/src/biblioteca/biblioteca/Util.java
@@ -1,0 +1,17 @@
+package biblioteca.biblioteca;
+
+import java.util.ArrayList;
+
+/**
+ *
+ * @author Luam
+ */
+public class Util {
+
+    public static void imprimir(ArrayList a) {
+        for (Object i : a) {
+            System.out.println(i);
+        }
+    }
+    
+}
